Frontend now sends position in HTTP headers
